@@ -1,24 +1,19 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.EntityFrameworkCore;
 using MVCStore.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace MVCStore.Data
 {
-    public class SeedData
+    public static class SeedData
     {
         public static void EnsurePopulated(IApplicationBuilder app)
         {
-            ApplicationDbContext context = app.ApplicationServices
-            .CreateScope().ServiceProvider.GetRequiredService<ApplicationDbContext>();
+            ApplicationDbContext context = app.ApplicationServices.CreateScope().ServiceProvider.GetRequiredService<ApplicationDbContext>();
+
             if (context.Database.GetPendingMigrations().Any())
             {
                 context.Database.Migrate();
             }
+
             if (!context.Products.Any())
             {
                 context.Products.AddRange(
@@ -42,9 +37,11 @@ namespace MVCStore.Data
                     Description = "FIFA-approved size and weight",
                     Category = "Soccer",
                     Price = 19.50m
-                });
+                }
+                );
+
+                context.SaveChanges();
             }
-            context.SaveChanges();
         }
     }
 }
