@@ -1,14 +1,11 @@
 ﻿using MVCStore.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace MVCStore.ViewModels
 {
     public class ProductsListViewModel
     {
         public IEnumerable<Product> Products { get; set; }
-        public PagingInfo PagingInfo { get; set; }
+        = Enumerable.Empty<Product>();
+        public PagingInfo PagingInfo { get; set; } = new();
     }
 }
