@@ -1,11 +1,9 @@
 ﻿@model ProductsListViewModel
-
-@foreach (var p in Model.Products)
+@foreach (var p in Model?.Products ?? Enumerable.Empty<Product>())
 {
     <partial name="ProductSummary" model="p" />
 }
-
-<div page-model="@Model.PagingInfo" page-action="Index" page-classes-enabled="true"
+<div page-model="@Model?.PagingInfo" page-action="Index" page-classes-enabled="true"
      page-class="btn" page-class-normal="btn-outline-dark"
      page-class-selected="btn-primary" class="btn-group pull-right m-1">
 </div>
