@@ -2,11 +2,12 @@
 <div class="card card-outline-primary m-1 p-1">
     <div class="bg-faded p-1">
         <h4>
-            @Model.Name
-            <span class="badge badge-pill badge-primary" style="float:right">
-                <small>@Model.Price.ToString("c")</small>
+            @Model?.Name
+            <span class="badge rounded-pill bg-primary text-white"
+                  style="float:right">
+                <small>@Model?.Price.ToString("c")</small>
             </span>
         </h4>
     </div>
-    <div class="card-text p-1">@Model.Description</div>
+    <div class="card-text p-1">@Model?.Description</div>
 </div>
